requirements.txt
anyio==3.6.2
azure-cognitiveservices-speech==1.28.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
google-api-core==2.11.0
google-auth==2.19.0
google-cloud-speech==2.19.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.14.0
httpcore==0.17.1
httpx==0.24.0
idna==3.4
MouseInfo==0.1.3
pafy==0.5.5
Pillow==9.5.0
proto-plus==1.22.2
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pygame==2.4.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.28
python-telegram-bot==20.3
pytube==15.0.0
pytweening==1.0.4
qrcode==7.4.2
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SpeechRecognition==3.10.0
typing_extensions==4.6.3
urllib3==1.26.16
